terrain/tile mode button switch fully implemented into map editor
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@130 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/Widget.cpp
@@ -49,7 +49,7 @@ void Widget::setSize(float inX, float inY)
 *   mouse.  Allows the widget to do things as the mouse
 *   moves around over it.
 **************************************************/
-void Widget::onMouseChange(float inX, float inY)
+void Widget::onMouseChange(int inX, int inY)
 {
 }
